max speed up with avx256

+++ src/max/max.hpp
@@ -10,8 +10,8 @@
 
 extern float rawFloatData[DATANUM];
 
-float Max(const float data[], const int len);
-float MaxSpeedUp(const float data[], const int len);
+float Max(const float data[], const size_t len);
+float MaxSpeedUp(const float data[], const size_t len);
 
 void InitData(const float data[], const size_t len);
